requirements.txt
aiofiles==24.1.0
alembic==1.14.0
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.30.0
blinker==1.6.3
buildpg==0.4
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.8
MarkupSafe==3.0.2
priority==2.0.0
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
pyhumps==3.8.0
python-dotenv==1.0.1
Quart==0.19.9
quart-auth==0.10.1
quart-cors==0.7.0
quart-db==0.9.0
quart-flask-patch==0.3.0
quart-rate-limiter==0.10.0
quart-schema==0.20.0
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
sqlapagination==0.0.1
tenacity==8.0.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.0.6
wsproto==1.2.0
WTForms==3.2.1